Make --html have geT_bugs.pl generate both txt and html versions
Since I become Release Maintainer (3.12.x) I was bothered by the fact that
generating the HTML release notes was a separate step. This might not be a
feature for others, so I won't complain if the next RM reverts this. But
for the 3.22 major release, I want it like this :-D

This patch makes a slight change in the behaviour of get_bugs.pl:
- It will now always create the text version of the release notes
- The --html switch will make get_bugs.pl generate both version in one
  run. I confess I was tempted to make it the default behaviour too.

Bye!