Stop using regexps to extract information from Pootle
This patch makes get_bugs.pl use HTML::TableExtract to better parse
Pootle's page. A proper solution should make use of Pootle's REST api
but it requires autentication for now, so in order to generate the
release notes for the 3.22 beta, this should be good enough.

The HTML::TableExtract object needs an arrayref of column names, to look
for the table that matches it, and returns the rows as arrays.

They are trimmed because of spaces and newlines, and then used in an
obvious way.

I remembered this 'historical' stackoverflow post that is related to this
issue :-D

http://stackoverflow.com/questions/1732348/regex-match-open-tags-except-xhtml-self-contained-tags

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>